Farther through refactor
